+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+### v1.1.0 (14/08/2021)
+* Added option to build the downloader as a native C file, greatly reduces detections
+* Added a TinyCC compiler for native C builds
+* Updated required .NET for the builder to .NET 4.5 and the required .NET for the managed .NET C# build to .NET 4.0
+* Changed the Run as Administrator option to use a new manifest file
 ### v1.0.0 (14/08/2021)
 * Initial release